+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_lights.py
@@ -42,6 +42,7 @@ class AllLightsInteractionPlugin(_StageManagerUSDInteractionPlugin):
 
     compatible_filters: list[str] = Field(
         default=[
+            "AdditionalFilterPlugin",
             "IgnorePrimsFilterPlugin",
             "IsCaptureFilterPlugin",
             "LightPrimsFilterPlugin",