+++ source/extensions/omni.flux.stage_manager.plugin.interaction.usd/omni/flux/stage_manager/plugin/interaction/usd/all_prims.py
@@ -33,12 +33,16 @@ class AllPrimsInteractionPlugin(_StageManagerUSDInteractionPlugin):
     )
     compatible_filters: list[str] = Field(
         default=[
+            "AdditionalFilterPlugin",
             "IgnorePrimsFilterPlugin",
             "IsCaptureFilterPlugin",
+            "IsCategoryFilterPlugin",
+            "MaterialPrimsFilterPlugin",
             "MeshPrimsFilterPlugin",
+            "LightPrimsFilterPlugin",
             "OmniPrimsFilterPlugin",
             "SearchFilterPlugin",
-            "IsCategoryFilterPlugin",
+            "SkeletonPrimsFilterPlugin",
         ],
         exclude=True,
     )