+++ source/extensions/lightspeed.trex.app.resources/docs/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.23.2]
+### Added
+- Added filter svg icons
+- Added `AdditionalFilterPlugin` to Stage Manager Schema
+
 ## [1.23.1]
 ### Changed
 - Hidden Edit/Preferences menu item for streamlined interface